alex/cdn-optimize: move
